revert to build only version edsl_doc.cc and edsl_doc.py edsl_doc.cc and edsl_doc.py moved to plaidml/edsl title changed back to welcome message code_examples folder deleted mote carlo pi example code removed
